Add support for bar chart exports into vector graphics.
